requirements.txt
openai>=1.0.0
fastmcp>=0.2.0
httpx>=0.24.0
python-dotenv>=1.0.0